requirements.txt
openai>=1.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
